refactor(cra-tests): #1556 Relocate GatewayApi endpoint definitions
Move `GatewayApi` from `gateway.py` to `http_api.py` to consolidate all HTTP endpoint paths, methods, statuses, and headers within a single module. `gateway.py` now re-exports `GatewayApi` for backward compatibility.

Enhance `EvidenceLog.finish()` to prefix all terminal log entries (UTC END, DURATION, OVERALL VERDICT) with the injected end timestamp and ensure the stream is flushed immediately before closing.
